Slight loop point adjustments + bug fix
Dire: Loop point was set 1 second too early
Scores: Cuts off a bit early, so loop point is 0.2 seconds before the end of the track now
Final: Really tiny adjustment

The bug was that -1 was interpreted as 1 second before the end of the track, when it's meant to be the VERY end of the track.
Also, there's a debug feature where you can hold L to fast forward through music tracks now.